refactor(core): simplify things a bit more
